web/requirements.txt
Flask==1.1.1
uWSGI==2.0.18
Flask-SQLAlchemy==2.4.1
psycopg2==2.8.4
Flask-Login==0.4.1
firebase-admin==3.2.1
grpcio==1.26.0
google-cloud-firestore==1.6.1
google-cloud-core==1.2.0